Move FooterNav into the Footer directory
